requirements.txt
# Community Volunteer Matcher Requirements
mcp[server] >= 1.0.0
descope >= 1.0.0
httpx >= 0.25.0
pydantic >= 2.0.0
uvicorn >= 0.24.0
python-dotenv >= 1.0.0
flask >= 2.0.0